Move Gateway Rejection Reason type and constants to their own file.
